requirements.txt
django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-filter==23.5
drf-yasg==1.21.7
django-cors-headers==4.3.1
django-restql==0.15.1
djangorestframework-simplejwt==5.1.0